dom/workers/test/test_fileReadSlice.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=664783
-->
<window title="Mozilla Bug 664783"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript" src="dom_worker_helper.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=664783"
     target="_blank">Mozilla Bug 664783</a>

  <div id="content" style="display: none">
    <input id="fileList" type="file"></input>
  </div>

  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  if (navigator.platform.startsWith("Win")) {
    SimpleTest.expectAssertions(0, 1);
  }

  /** Test for Bug 664783 **/

  var fileNum = 0;

  /**
   * Create a file which contains the given data.
   */
  function createFileWithData(fileData) {
    var testFile = Components.classes["@mozilla.org/file/directory_service;1"]
                       .getService(Components.interfaces.nsIProperties)
                       .get("ProfD", Components.interfaces.nsIFile);
    testFile.append("workerReadSlice" + fileNum++);

    var outStream = Components.classes["@mozilla.org/network/file-output-stream;1"]
                        .createInstance(Components.interfaces.nsIFileOutputStream);
    outStream.init(testFile, 0x02 | 0x08 | 0x20, // write, create, truncate
                   0666, 0);
    outStream.write(fileData, fileData.length);
    outStream.close();

    var fileList = document.getElementById('fileList');
    fileList.value = testFile.path;

    return fileList.files[0];
  }

  /**
   * Creates a worker which slices a blob to the given start and end offset and
   * reads the content as text.
   */
  function readSlice(blob, start, end, expectedText) {
    var worker = new Worker("fileReadSlice_worker.js");

    worker.onerror = function(event) {
      ok(false, "Worker had an error: " + event.message);
      finish();
    };

    worker.onmessage = function(event) {
      is(event.data, expectedText, "Text from sliced blob in worker is incorrect.");
      finish();
    };

    var params = {blob: blob, start: start, end: end};
    worker.postMessage(params);
    waitForWorkerFinish();
  }

  // Empty file.
  readSlice(createFileWithData(""), 0, 0, "");

  // Typical use case.
  readSlice(createFileWithData("HelloBye"), 5, 8, "Bye");

  // End offset too large.
  readSlice(createFileWithData("HelloBye"), 5, 9, "Bye");

  // Start of file.
  readSlice(createFileWithData("HelloBye"), 0, 5, "Hello");

  ]]>
  </script>
</window>
